pyproject.toml
[tool.poetry]
name = "langchain-demo"
version = "0.1.0"
description = ""
authors = ["Subodh Jena <jenasubodh@gmail.com>"]
readme = "README.md"
packages = [{include = "langchain_demo"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
streamlit = "^1.21.0"
langchain = "^0.0.150"
openai = "^0.27.4"
wikipedia = "^1.4.0"
tiktoken = "^0.3.3"
chromadb = "^0.3.21"
unstructured = "^0.6.2"
pypdf = "^3.8.1"
deeplake = "^3.3.2"
playwright = "^1.32.1"
pdfminer-six = "^20221105"
google-api-python-client = "^2.86.0"
selenium = "^4.9.0"
pyperclip = "^1.8.2"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"
watchdog = "^3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
